Renamings
- Move BuildFrom to a separate file
- Rename CanBuild to Factory
- Add more scaladoc comments
