+++ src/main/scala/strawman/collection/Iterator.scala
@@ -752,10 +752,10 @@ trait Iterator[+A] extends IterableOnce[A] { self =>
   /** Converts this Iterator into another collection.
     *  @return a new collection containing all elements of this Iterator.
     *  @tparam C The collection type to build.
-    *  @param canBuild Collection factory to use. The factory may or may
+    *  @param factory Collection factory to use. The factory may or may
     *                  not eagerly consume this iterator.
     */
-  def to[C](canBuild: CanBuild[A, C]): C = canBuild.fromSpecific(self)
+  def to[C](factory: Factory[A, C]): C = factory.fromSpecific(self)
 
 }
